Move normalization code from Dhall.Core to a new module D.Normalize
Closes #1127.
